refactor electrostatic
