+++ adept/tf1d/solvers/pushers.py
@@ -5,7 +5,7 @@ from jax import numpy as jnp
 import numpy as np
 import equinox as eqx
 
-from adept.theory.electrostatic import get_complex_frequency_table
+from adept.electrostatic import get_complex_frequency_table
 from adept import get_envelope
