+++ tests/test_tf1d/test_landau_damping.py
@@ -11,8 +11,7 @@ config.update("jax_enable_x64", True)
 from jax import numpy as jnp
 import mlflow
 
-from adept.theory import electrostatic
-from adept import ergoExo
+from adept import ergoExo, electrostatic
 
 
 def _modify_defaults_(defaults, rng):
