+++ tests/test_tf1d/test_resonance.py
@@ -9,10 +9,7 @@ config.update("jax_enable_x64", True)
 # config.update("jax_disable_jit", True)
 
 from jax import numpy as jnp
-import mlflow
-
-from adept.theory import electrostatic
-from adept import ergoExo
+from adept import ergoExo, electrostatic
 
 
 def _modify_defaults_(defaults, rng, gamma):
