+++ tests/test_vlasov1d/test_landau_damping.py
@@ -10,8 +10,7 @@ from jax import config
 config.update("jax_enable_x64", True)
 # config.update("jax_disable_jit", True)
 
-from adept.theory import electrostatic
-from adept import ergoExo
+from adept import ergoExo, electrostatic
 
 
 def _modify_defaults_(defaults, rng, real_or_imag, time, field, edfdv):
